web_service/requirements.txt
fastapi
uvicorn
numpy
requests
Pillow
google-generativeai
textwrap3
IPython